% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DALY.R
\name{calculate_DALY}
\alias{calculate_DALY}
\title{Calculate disbility-adjusted life years (DALY)}
\usage{
calculate_DALY(
  disease,
  country,
  cases,
  year = NULL,
  life_expectancy = NULL,
  disability_weight_severe = NULL,
  disability_weight_moderate = NULL,
  disability_weight_mild = NULL,
  prob_inpatient = NULL,
  prob_outpatient = NULL,
  prob_nocare = NULL,
  case_fatality_inpatient = NULL,
  case_fatality_outpatient = NULL,
  case_fatality_nocare = NULL,
  illness_duration = NULL
)
}
\arguments{
\item{year}{A range of years during which}

\item{life_expectancy}{Life expectancy at the time of death in years}

\item{illness_duration}{The duration of the disease in terms of days}

\item{country_name}{The of country of interest for which you want to set up the population in}

\item{population_data}{The population data of countries which includes the file}

\item{disability_weight}{A fraction that expresses the severity of the disease compared to the death}
}
\description{
The \code{calculate_DALY()} is used to calculate DALY and a wrapper of \code{setup_cohorts()}, \code{calculate_cases()}, \code{calculate_YLL()} and \code{calculate_YLD()} functions
}
\examples{
compute_DALY()
}
