% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vaccine_protected.R
\name{calculate_vaccine_protected}
\alias{calculate_vaccine_protected}
\title{Vaccinate population for a set of condition}
\usage{
calculate_vaccine_protected(
  disease = NULL,
  country = NULL,
  population = NULL,
  year = 2000:2100,
  vaccine_coverage = NULL,
  vaccine_efficacy = NULL,
  vaccine_immunity_duration = NULL,
  exponential_decay = TRUE,
  vaccine_efficacy_by_year = NULL
)
}
\arguments{
\item{year}{Year in which vaccination is implemented}

\item{efficacy}{Vaccine efficacy as a proportion}

\item{duration_effect}{Years during which vaccine-induced immunity remains}

\item{coverage}{A proportion of target population that receives the vaccine}

\item{age_range}{Target age range}
}
\description{
Vaccinate population for a set of condition
}
\examples{
pop <- setup_population(); compute_cases(pop)
}
