% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_YLD.R
\name{calculate_YLD}
\alias{calculate_YLD}
\title{Calculate years lost due to disability (YLD)}
\usage{
calculate_YLD(
  disease = NULL,
  cases = NULL,
  disability_weight_severe = NULL,
  disability_weight_moderate = NULL,
  disability_weight_mild = NULL,
  prob_inpatient = NULL,
  prob_outpatient = NULL,
  prob_nocare = NULL,
  case_fatality_inpatient = NULL,
  case_fatality_outpatient = NULL,
  case_fatality_nocare = NULL,
  illness_duration = NULL
)
}
\arguments{
\item{disease}{The name of disease of interest. Now it only supports Cholera or Typhoid}

\item{cases}{The number of cases with diseases of interest}

\item{illness_duration}{Duration of illness in terms of year}

\item{disability_weight}{A fraction that expresses the severity of the disease compared to the death}
}
\description{
The \code{calculate_YLD()} computes YLD.
}
\examples{
compute_YLD(data.frame(a=100), 0.2, 12/365)
}
