% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_YLL.R
\name{calculate_YLL}
\alias{calculate_YLL}
\title{Calculate years of life lost (YLL)}
\usage{
calculate_YLL(
  disease = c("Typhoid", "Cholera"),
  cases = NULL,
  country = NULL,
  year = NULL,
  life_expectancy = NULL,
  prob_inpatient = NULL,
  prob_outpatient = NULL,
  prob_nocare = NULL,
  case_fatality_inpatient = NULL,
  case_fatality_outpatient = NULL,
  case_fatality_nocare = NULL
)
}
\arguments{
\item{life_expectancy}{Life expectancy at the time of death in years}

\item{deaths}{The number of deaths due to the disease of interest}
}
\description{
The \code{calculate_YLL()} calculatess YLL
}
\examples{
calculate_YLL(cases = cases)
}
