% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_country_names.R
\name{clean_country_names}
\alias{clean_country_names}
\title{Clean and standardize country names according to VIMC report templates}
\usage{
clean_country_names(country)
}
\arguments{
\item{country}{A vector of country names in character}
}
\description{
The \code{clean_country_names()} is used to clean and standardize country names according to VIMC report templates
population for both sexes and incidence rate
}
\examples{
country <- clean_country_names(country = "DRC")
# Congo, the Democratic Republic of the
}
