% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gavi201910_int_pop_both}
\alias{gavi201910_int_pop_both}
\title{Population size by age (0-120) and year (1885-2100).}
\format{
A data.table and data.frame:	19695 obs. of  222 variables:
\describe{
  \item{country }{country name}
  \item{age_from}{starting age}
  \item{age_to}{ending age}
  \item{gender}{gender}
  ...
}
}
\source{
\url{https://population.un.org/wpp/}
}
\usage{
gavi201910_int_pop_both
}
\description{
A dataset containing the population size (UN projections) by one-year age group up to 100 and 100-120 yo and country.
}
\keyword{datasets}
