% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso3.R
\name{get_iso3}
\alias{get_iso3}
\title{Get ISO3 codes for country names}
\usage{
get_iso3(country, lookup_table)
}
\arguments{
\item{country}{The input country names}
}
\description{
The \code{get_iso3()} is used to retrieve ISO3 codes for the country name of interest
population for both sexes and incidence rate
}
\examples{
iso3 <- get_iso3(country = "DRC")
# COd
}
