% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cohorts.R
\name{setup_cohorts}
\alias{setup_cohorts}
\title{Base population setup}
\usage{
setup_cohorts(country = NULL, year = NULL, population_data = NULL)
}
\arguments{
\item{country}{The of country of interest for which you want to set up the population in}

\item{year}{A range of years during which cohorts are created}

\item{sanitation_RR}{Relative risk of those with at least basic sanitation according to JMP WASH definition}
}
\description{
The \code{setup_population ()} sets up cohorts using the interpolated population for both sexes
age distribution in 2020
}
\examples{
setup_population()

}
