% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoch_run_scenario.R,
%   R/stoch_run_scenario_archive.R
\name{stoch_run_scenario}
\alias{stoch_run_scenario}
\title{Run models to compute cohort size, cases, deaths, dalys
The \code{stoch_run_scenario()} runs runs for a poisson-distributed incidence rate, I think this needs to be expanded to include other variables}
\usage{
stoch_run_scenario(
  disease = NULL,
  country = NULL,
  vacc_scenario = NULL,
  runs = 30,
  year = 2000:2100
)

stoch_run_scenario(
  disease = NULL,
  country = NULL,
  vacc_scenario = NULL,
  runs = 30,
  year = 2000:2100
)
}
\arguments{
\item{disease}{Target disease (i.e., cholera or typhoid)}

\item{country}{name of the country}

\item{year}{Year in which vaccination is implemented}

\item{vacc_input}{vaccine coverage information in the long format}
}
\description{
Run models to compute cohort size, cases, deaths, dalys
The \code{stoch_run_scenario()} runs runs for a poisson-distributed incidence rate, I think this needs to be expanded to include other variables

Vaccinate population for a set of condition
}
\examples{
pop <- stoch_run_scenario(disease = Typhoid, country = target_countries, vacc_scenario = "campaign");
pop <- setup_population(); compute_cases(pop)
}
