% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_val_plot.R
\name{year_val_plot}
\alias{year_val_plot}
\title{Ploting year vs. variable values}
\usage{
year_val_plot(
  disease = NULL,
  country = NULL,
  vacc_scenario = NULL,
  sim_central_value,
  age
)
}
\arguments{
\item{disease}{disease (i.e., either Typhoid or Cholera)}

\item{country}{country of interest in long name format}

\item{age}{age set to some value if you want to see the results for a particular age. If missing, results are summed across ages}

\item{x}{VIMC long-format result}
}
\description{
The \code{year_val_plot()} filiters and plots the value of interest
}
\examples{
year_val_plot()

}
