% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DALY.R
\name{calculate_DALY}
\alias{calculate_DALY}
\title{Calculate disbility-adjusted life years (DALY)}
\usage{
calculate_DALY(
  disease = NULL,
  country = NULL,
  cases = NULL,
  year = 2000:2100,
  disability_weight = NULL,
  illness_duration = NULL,
  case_fatality = NULL,
  life_expectancy_data = "data/201910gavi-5_dds-201910_2_life_ex_both.csv"
)
}
\arguments{
\item{year}{A range of years during which}

\item{disability_weight}{A fraction that expresses the severity of the disease compared to the death}

\item{illness_duration}{The duration of the disease in terms of days}

\item{country_name}{The of country of interest for which you want to set up the population in}

\item{population_data}{The population data of countries which includes the file}

\item{life_expectancy}{Life expectancy at the time of death in years}
}
\description{
The \code{calculate_DALY()} is used to calculate DALY and a wrapper of \code{setup_cohorts()}, \code{calculate_cases()}, \code{calculate_YLL()} and \code{calculate_YLD()} functions
}
\examples{
compute_DALY()
}
