% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_YLL.R
\name{calculate_YLL}
\alias{calculate_YLL}
\title{Calculate years of life lost (YLL)}
\usage{
calculate_YLL(
  cases = NULL,
  disease = c("Typhoid", "Cholera"),
  country = NULL,
  year = 2000:2100,
  case_fatality = NULL,
  life_expectancy_data = "data/201910gavi-5_dds-201910_2_life_ex_both.csv"
)
}
\arguments{
\item{deaths}{The number of deaths due to the disease of interest}

\item{life_expectancy}{Life expectancy at the time of death in years}
}
\description{
The \code{calculate_YLL()} calculatess YLL
}
\examples{
calculate_YLL(cases = cases)
}
