% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vaccine_coverage.R
\name{create_vaccine_coverage}
\alias{create_vaccine_coverage}
\title{Vaccinate population for a set of condition}
\usage{
create_vaccine_coverage(
  disease = NULL,
  country = NULL,
  vacc_input = NULL,
  routine = TRUE,
  num_row = 101,
  num_col = 101,
  year = 2000:2100
)
}
\arguments{
\item{country}{name of the country}

\item{vacc_input}{vaccine coverage information in the long format}

\item{year}{Year in which vaccination is implemented}
}
\description{
Vaccinate population for a set of condition
}
\examples{
pop <- setup_population(); compute_cases(pop)
}
