% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoch_run_scenario.R
\name{stoch_run_scenario}
\alias{stoch_run_scenario}
\title{Vaccinate population for a set of condition}
\usage{
stoch_run_scenario(
  disease = NULL,
  country = NULL,
  vacc_scenario = NULL,
  runs = 30,
  year = 2000:2100
)
}
\arguments{
\item{country}{name of the country}

\item{year}{Year in which vaccination is implemented}

\item{vacc_input}{vaccine coverage information in the long format}
}
\description{
Vaccinate population for a set of condition
}
\examples{
pop <- setup_population(); compute_cases(pop)
}
