% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimc_report.R
\name{vimc_report}
\alias{vimc_report}
\title{VIMC report}
\usage{
vimc_report(
  disease = NULL,
  country = NULL,
  sim_output,
  year = 2000:2100,
  age = 0:100,
  value_name
)
}
\arguments{
\item{sim_output}{The population data of countries which includes the file}

\item{template}{The VIMC template used}
}
\description{
The \code{vimc_report()} reshapes simulation results (e.g., data frame in the wide format or a list of wide-format data frames into the long format
}
\examples{
vimc_report()

}
