% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimc_stoch_report.R
\name{vimc_stoch_report}
\alias{vimc_stoch_report}
\title{VIMC report for individual runs from stochastic simulation}
\usage{
vimc_stoch_report(
  disease = NULL,
  country = NULL,
  sim_output,
  year = 2000:2100,
  age = 0:100,
  value_name,
  column_order = c("disease", "run_id", "year", "age", "country", "country_name",
    "cohort_size", "cases", "deaths", "dalys")
)
}
\arguments{
\item{sim_output}{The population data of countries which includes the file}

\item{template}{The VIMC template used}
}
\description{
The \code{vimc_stoch_report()} reshapes simulation results (e.g., data frame in the wide format or a list of wide-format data frames into the long format
}
\examples{
vimc_stoch_report()

}
